snapshots: move DAL code and tests
