+++ silkworm/db/bodies/body_snapshot.cpp
