+++ silkworm/db/snapshot_index_builder_test.cpp
@@ -14,15 +14,14 @@
    limitations under the License.
 */
 
-#include "index_builder.hpp"
-
 #include <catch2/catch.hpp>
 
-#include <silkworm/db/snapshots/body_index.hpp>
-#include <silkworm/db/snapshots/header_index.hpp>
-#include <silkworm/db/snapshots/test_util/common.hpp>
-#include <silkworm/db/snapshots/txn_index.hpp>
-#include <silkworm/db/snapshots/txn_to_block_index.hpp>
+#include <silkworm/db/bodies/body_index.hpp>
+#include <silkworm/db/headers/header_index.hpp>
+#include <silkworm/db/snapshots/index_builder.hpp>
+#include <silkworm/db/test_util/temp_snapshots.hpp>
+#include <silkworm/db/transactions/txn_index.hpp>
+#include <silkworm/db/transactions/txn_to_block_index.hpp>
 #include <silkworm/infra/common/decoding_exception.hpp>
 #include <silkworm/infra/common/directories.hpp>
 #include <silkworm/infra/test_util/log.hpp>
