+++ silkworm/db/snapshot_sync.cpp
@@ -23,9 +23,9 @@
 #include <magic_enum.hpp>
 
 #include <silkworm/core/types/hash.hpp>
+#include <silkworm/db/headers/header_snapshot.hpp>
 #include <silkworm/db/mdbx/etl_mdbx_collector.hpp>
 #include <silkworm/db/snapshots/config.hpp>
-#include <silkworm/db/snapshots/header_snapshot.hpp>
 #include <silkworm/db/snapshots/index_builder.hpp>
 #include <silkworm/db/snapshots/path.hpp>
 #include <silkworm/db/stages.hpp>
