+++ silkworm/db/transactions/txn_index.cpp
@@ -16,8 +16,9 @@
 
 #include "txn_index.hpp"
 
-#include "body_txs_amount_query.hpp"
-#include "snapshot_reader.hpp"
+#include <silkworm/db/bodies/body_txs_amount_query.hpp>
+#include <silkworm/db/snapshots/snapshot_reader.hpp>
+
 #include "txn_snapshot_word_serializer.hpp"
 
 namespace silkworm::snapshots {
