+++ silkworm/db/transactions/txs_and_bodies_query.hpp
@@ -26,10 +26,9 @@
 #include <silkworm/core/common/base.hpp>
 #include <silkworm/core/common/bytes.hpp>
 #include <silkworm/core/types/block_body_for_storage.hpp>
+#include <silkworm/db/snapshots/path.hpp>
 #include <silkworm/db/snapshots/seg/decompressor.hpp>
 #include <silkworm/infra/common/memory_mapped_file.hpp>
-
-#include "path.hpp"
 
 namespace silkworm::snapshots {
